Function to replace population with elitism and diversity preservation

Function to replace population with elitism and diversity preservation

Usage

replacement_cpp(
  population,
  offspring,
  fitness,
  offspring_fitness,
  num_to_replace
)

Arguments

population

Numeric matrix of individuals.

offspring

Numeric matrix of offspring.

fitness

Numeric matrix of population fitness scores.

offspring_fitness

Numeric matrix of offspring fitness scores.

num_to_replace

Number of individuals to replace.

Value

Numeric matrix of updated population.

Examples

genomic_data <- matrix(rnorm(100), nrow = 10, ncol = 10)
population <- BioGA::initialize_population_cpp(genomic_data, 
population_size = 5)
fitness <- BioGA::evaluate_fitness_cpp(genomic_data, population, 
c(1.0, 0.5))
selected_parents <- BioGA::selection_cpp(population, fitness, 
num_parents = 2)
#> Current front size: 1
offspring <- BioGA::crossover_cpp(selected_parents, offspring_size = 2)
offspring_fitness <- BioGA::evaluate_fitness_cpp(genomic_data, offspring, 
c(1.0, 0.5))
BioGA::replacement_cpp(population, offspring, fitness, offspring_fitness, 
num_to_replace = 1)
#>            [,1]       [,2]       [,3]       [,4]        [,5]       [,6]
#> [1,] -0.7526890 -1.7267304 -0.2656516  0.8496430  1.90236182 -0.1009749
#> [2,]  1.0541810 -0.3439172  1.3001987  2.0024827 -0.08856511 -0.1009749
#> [3,] -0.5021987 -1.3387743 -1.3952743 -0.4371595  0.06670087  0.1748027
#> [4,] -0.7526890 -1.7267304 -0.2656516  0.8496430  1.90236182 -0.1009749
#> [5,] -0.7526890 -1.7267304 -0.2656516  0.8496430  1.90236182 -0.1009749
#>            [,7]       [,8]       [,9]      [,10]
#> [1,] -0.2007810 -0.1136399 0.16806538  0.2374303
#> [2,]  1.6408462  0.4281668 0.00729009  0.8450130
#> [3,]  0.6307541 -0.2190504 2.03757402 -0.7152422
#> [4,] -0.2007810 -0.1136399 0.16806538  0.2374303
#> [5,] -0.2007810 -0.1136399 0.16806538  0.2374303